refactor: Move paragraph to data
p 태그 내 데이터를 분리하여 각 data 파일로 옮겼습니다.
