+++ src/pages/TermsPage.jsx
@@ -1,5 +1,6 @@
 import { useEffect } from "react";
 import BackButtonHeader from "../components/common/BackButtonHeader";
+import TermsData from "../components/common/TermsData";
 
 const TermsPage = () => {
   useEffect(() => {
@@ -14,378 +15,7 @@ const TermsPage = () => {
         <span className="text-sm pl-10">이용약관</span>
       </BackButtonHeader>
       <div className="p-[10px]">
-        <p>
-          제 1 장 총칙
-          <br />
-          제1조 (목적)
-          <br />
-          1. 이 약관은 순수웨딩(이하 &apos;회사&apos;)이 제공하는
-          https://sunsu.user-app.krampoline.com(이하 &apos;서비스&apos;)의
-          이용조건 및 절차에 관한 사항을 규정함을 목적으로 합니다.
-          <br />
-          제2조 (정의)
-          <br />
-          1. 이 약관에서 사용하는 용어는 다음과 같습니다..
-          <br />
-          가. &apos;회원&apos;이라 함은 서비스를 제공받기 위해 회사가 인정하는
-          절차를 통해 가입하여 이용자번호(ID)를 부여 받은 자를 말합니다.
-          <br />
-          나. &apos;이용자번호(ID)&apos;라 함은 가입회원의 식별과 회원의 서비스
-          이용을 위해 고객이 선정하고 회사가 부여하는 문자와 숫자의 조합을
-          말합니다.
-          <br />
-          다. &apos;비밀번호&apos;라 함은 회원이 부여받은 이용자번호(ID)와
-          일치된 회원임을 확인하고 회원의 권익보호를 위하여 회원이 선정한 문자와
-          숫자의 조합을 말합니다.
-          <br />
-          2. 이 약관에서 사용하는 용어의 정의는 제1항에서 정하는 것을 제외하고는
-          관계 법령 및 서비스별 이용약정에서 정하는 바에 의합니다.
-          <br />
-          제3조 (약관의 효력 및 변경)
-          <br />
-          1. 회사는 약관의 규제에 관한 법률, 전자상거래등에서의 소비자보호에
-          관한 법률, 정보통신망이용촉진 및 정보보호 등에 관한 법률 등 관련법을
-          위배하지 않는 범위에서 본 약관을 개정할 수 있습니다.
-          <br />
-          2. 회사는 본 약관을 개정할 경우에는 적용일자 및 개정사유를 명시하여
-          현행 약관과 함께 회사가 제공하는 서비스 사이트의 초기 화면에 그
-          적용일자 7일 이전부터 적용일 후 상당기간 동안 공지하고 기존 회원에게는
-          개정약관을 전자우편 주소로 발송합니다. 다만, 회원에게 불리하게
-          약관내용을 변경하는 경우에는 최소한 30일 이상의 사전 유예기간을 두고
-          공지합니다. 이 경우 회사는 개정 전 내용과 개정 후 내용을 명확하게
-          비교하여 회원이 알기 쉽도록 표시합니다.
-          <br />
-          3. 회사가 약관을 개정할 경우에는 개정약관 공지 후 개정약관의 적용에
-          대한 이용자의 동의여부를 확인합니다. 이용자가 개정약관의 적용에
-          동의하지 않는 경우 회사 또는 이용자는 콘텐츠 이용계약을 해지할 수
-          있습니다. 다만, 동의기간 내에 거절의 의사를 표시하지 않으면 동의한
-          것으로 본다는 내용을 별도로 고지한 경우, 이용자가 이에 따른 거절의
-          의사표시를 하지 않았을 때에는 변경약관에 동의한 것으로 간주합니다.
-          <br />
-          제4조 (약관외 적용범위)
-          <br />
-          1. 이 약관에 명시되지 않은 사항은 관계 법령에 규정되어 있을 경우 그
-          규정에 따르며, 그렇지 않은 경우에는 정부가 제정한
-          전자거래소비자보호지침 및 관계법령 또는 일반적인 관례에 따릅니다.
-          <br />
-          <br />
-          제 2 장 이용계약
-          <br />
-          제 5조 (이용계약의 성립)
-          <br />
-          1. 이용계약은 본 약관에 동의한 고객이 회사 소정의 이용신청서를 온라인
-          또는 회사가 정하는 방법으로 신청, 회사의 승낙에 의하여 성립합니다.
-          이용계약은 이용자번호(ID) 단위로 체결됩니다.
-          <br />
-          2. 제 1항의 규정에 의해 고객이 이용신청을 할 때에는 회사가 정한
-          필수입력 사항을 입력하여야 합니다.
-          <br />
-          3. 회사는 고객의 이용신청이 다음 각 호에 해당하는 경우 그 신청에 대한
-          승낙을 제한 또는 유보할 수 있습니다.
-          <br />
-          가. 성명이 실명이 아닐 경우
-          <br />
-          다. 선량한 풍속, 기타 사회질서를 해 할 목적으로 신청한 경우
-          <br />
-          라. 이용신청서의 내용을 허위로 기재하였거나 허위서류를 첨부하였을 경우
-          <br />
-          마. 회사의 업무수행상 또는 기술상 지장이 있는 경우
-          <br />
-          바. 기타 이용신청고객의 귀책사유로 이용승낙이 곤란한 경우
-          <br />
-          제 6조 (서비스 이용 신청)
-          <br />
-          1. 회원으로 가입하여 본 서비스를 이용하고자 하는 이용고객은 회사에서
-          요청하는 제반정보 (이름, 전자우편, 휴대폰 번호 등)를 제공하여야
-          합니다.
-          <br />
-          2. 모든 회원은 반드시 회원 본인의 이름과 전자우편 그리고 휴대폰 번호를
-          제공하여야만 서비스를 이용할 수 있으며, 실명으로 등록하지 않은
-          사용자는 일체의 권리를 주장할 수 없습니다.
-          <br />
-          3. 회원가입은 반드시 실명으로만 가입할 수 있으며 회사는 실명확인조치를
-          할 수 있습니다.
-          <br />
-          4. 타인의 명의(이름 및 전자우편)를 도용하여 이용신청을 한 회원의 모든
-          ID는 삭제되며, 관계법령에 따라 처벌을 받을 수 있습니다.
-          <br />
-          5. 회사는 본 서비스를 이용하는 회원에 대하여 등급별로 제공되는
-          서비스에 차등을 둘 수 있습니다.
-          <br />
-          제7조 이용 신청의 승낙과 제한
-          <br />
-          1. 회사는 회원이 제6조에서 정한 모든 사항을 정확하게 기재하여
-          이용신청을 하였을 때 특별한 사정이 없는 한 접수 순서대로 이용신청을
-          승낙합니다.
-          <br />
-          2. 회사는 다음 각 호에 해당하는 이용신청에 대하여는 승낙을 유보할 수
-          있습니다.
-          <br />
-          가. 설비에 여유가 없는 경우
-          <br />
-          나. 기술상 지장이 있는 경우
-          <br />
-          다. 기타 회사의 사정상 이용승낙이 곤란한 경우
-          <br />
-          3. 회사는 다음 각 호에 해당하는 이용신청에 대하여는 이를 승낙하지 아니
-          할 수 있습니다.
-          <br />
-          가. 본인의 실명으로 신청하지 않은 경우
-          <br />
-          나. 다른 사람의 명의를 사용하여 신청한 경우
-          <br />
-          다. 이용 신청 시 필요내용을 허위로 기재하였거나 허위서류를 첨부하여
-          신청한 경우
-          <br />
-          라. 사회의 안녕질서 또는 미풍양속을 저해할 목적으로 신청한 경우
-          <br />
-          마. 기타 회사가 정한 이용신청요건이 미비되었을 때
-          <br />
-          바. 관계법령 위반 시
-          <br />
-          제8조 (회원탈퇴 및 자격 상실 등)
-          <br />
-          1. 회원은 회사에 언제든지 탈퇴를 요청할 수 있으며, 회사는 즉시
-          회원탈퇴를 처리합니다. 단 탈퇴한 회원은 동일한 아이디로 회원가입을 할
-          수 없습니다.
-          <br />
-          2. 회원이 다음 각호의 사유에 해당하는 경우, 회사는 회원자격을 제한 및
-          정지시킬 수 있습니다.
-          <br />
-          가. 가입신청 시에 허위내용을 등록한 경우
-          <br />
-          나. 회사의 서비스이용대금, 기타 회사의 서비스이용에 관련하여 회원이
-          부담하는 채무를 기일에 이행하지 않는 경우
-          <br />
-          다. 다른 사람의 회사의 서비스이용을 방해하거나 그 정보를 도용하는 등
-          전자상거래 질서를 위협하는 경우
-          <br />
-          라. 회사를 이용하여 법령 또는 본 약관이 금지하거나 공서양속에 반하는
-          행위를 하는 경우
-          <br />
-          3. 회사가 회원자격을 제한 , 정지시킨 후, 동일한 행위가 2회 이상
-          반복되거나 30일 이내에 그 사유가 시정되지 아니하는 경우 회사는
-          회원자격을 상실시킬 수 있습니다.
-          <br />
-          4. 회사가 회원자격을 상실시키는 경우에는 회원등록을 말소합니다. 이
-          경우 회원에게 이를 통지하고, 회원등록 말소 전에 최소한 30일 이상의
-          기간을 정하여 소명할 기회를 부여합니다.
-          <br />
-          제9조 (약관의 동의)
-          <br />
-          1. 회원이 등록절차를 거처 동의 버튼을 누름으로써 본 약관에 동의한
-          것으로 간주합니다.
-          <br />
-          <br />
-          제 3 장 계약당사자의 의무
-          <br />
-          제 10조 (회사의 의무)
-          <br />
-          1. 회사는 법령 및 본 약관에서 금하는 행위를 하지 않으며, 지속적이고
-          안정적으로 서비스를 제공하기 위해서 노력합니다.
-          <br />
-          2. 회사는 서비스 제공과 관련해 알고 있는 회원의 개인정보를 본인의
-          승낙없이 제3자에게 누설하거나 배포하지 않습니다. 단, 법률에 의한
-          국가기관의 요구나 수사상의 목적, 개인을 식별할 수 없는 인구통계학적
-          자료로는 제공될 수 있습니다.
-          <br />
-          3. 회사는 회원이 제기하는 의견이나 불만이 정당하다고 인정되면 즉시
-          처리합니다. 이용자가 제기한 의견이나 불만사항에 대해서는 게시판을
-          활용하거나 전자우편 등을 통하여 그 처리과정 및 결과를 전달합니다.
-          다만, 즉시 처리가 곤란할 경우에는 회원에게 그 사유와 일정을 통보하여야
-          합니다.
-          <br />
-          4. 회사는 회원이 안전하게 서비스를 이용할 수 있도록 회원의 개인정보 및
-          신용정보 보호를 위한 보안시스템을 구축합니다.
-          <br />
-          5. 회사는 지속적이고 안정적인 서비스를 제공하기 위하여 설비에 장애가
-          있거나 망실된 때에는 이를 지체없이 수리 또는 복구합니다. 다만,
-          천재지변이나 비상사태, 기타 부득이한 경우에는 그 서비스를 일시
-          중단하거나 중지할 수 있습니다. 이때, 회사는 지체없이 회원에게
-          통보하여야 합니다.
-          <br />
-          제 11조 (회원의 의무)
-          <br />
-          1. 회원 가입시에 입력하는 회원의 정보는 정확하여야 합니다. 또한, 이미
-          제공된 회원정보가 정확한 정보가 되도록 유지, 갱신하여야 합니다.
-          <br />
-          2. 회원은 공공의 안녕질서 및 미풍양속을 저해하는 다음 각 호의 통신
-          활동을 하여서는 안됩니다.
-          <br />
-          가. 범죄행위를 목적으로 하거나 범죄행위를 교사하는 내용
-          <br />
-          나. 선량한 풍속, 기타 사회질서를 해하는 내용
-          <br />
-          다. 타인의 명예를 손상시키거나 불이익을 주는 행위
-          <br />
-          라. 타인의 이용자번호(ID) 및 비밀번호를 사용하는 행위
-          <br />
-          마. 다른 회원 또는 제3자의 지적재산권을 침해하는 행위
-          <br />
-          바. 정보통신설비의 오동작이나 정보 등의 파괴 및 혼란을 유발시키는
-          컴퓨터 바이러스, 바이러스 감염자료를 등록 또는 유포하는 행위
-          <br />
-          사. 회사의 승인을 받지 않은 광고, 판촉물, 스팸메일, 행운의 편지,
-          피라미드 조직, 기타 다른 형태의 권유를 게시, 게재, 메일링 또는 다른
-          방법으로 전송하는 경우
-          <br />
-          아. 서비스를 통하여 얻은 정보를 회사의 사전 승락없이 복제 또는
-          유통시키거나 상업적으로 이용하는 경우
-          <br />
-          자. 기타 법령에 위배되는 행위
-          <br />
-          3. 이용자번호(ID)와 비밀번호(8~20자, 영문/숫자/특수문자 혼용, 단순문자
-          또는 개인정보 사용금지)의 관리책임은 회원 본인에게 있으며 비밀번호가
-          타인에게 노출된 경우 즉시 변경하여야 합니다.
-          <br />
-          4. 회원에게 부여된 이용자번호(ID)와 비밀번호의 관리소홀, 부정사용에
-          의하여 발생하는 모든 결과에 대한 책임은 회원에게 있습니다. 본인의
-          이용자번호(ID)가 부정하게 사용된 경우, 회원은 반드시 회사에 그 사실을
-          통보해야 합니다.
-          <br />
-          5. 회원은 이 약관 및 관계법령을 준수하여야 합니다.
-          <br />
-          6. 회사는 회원이 상기 제 조항을 위반했을 경우 즉시 계약해지 및
-          이용제한을 할 수 있고 이로 인한 손해배상을 청구할 수 있습니다
-          <br />
-          <br />
-          제 4 장 서비스의 이용
-          <br />
-          제 12조 서비스의 이용 개시
-          <br />
-          1. 회사는 회원의 이용 신청을 승낙한 때부터 서비스를 개시합니다. 단,
-          일부 서비스의 경우에는 지정된 일자부터 서비스를 개시합니다.
-          <br />
-          2. 회사의 업무상 또는 기술상의 장애로 인하여 서비스를 개시하지 못하는
-          경우에는 사이트에 공시하거나 회원에게 이를 통지합니다.
-          <br />
-          제 13조 (게시물 삭제)
-          <br />
-          1. 회사는 회원이 게시하거나 제공하는 서비스 내용이 다음 각 호에
-          해당한다고 판단될 경우 사전 통지없이 삭제할 수 있습니다.
-          <br />
-          가. 다른 회원 또는 제3자를 비방하거나 명예를 손상시키는 내용인 경우
-          <br />
-          나. 허위 사실의 유포
-          <br />
-          다. 공공질서 및 미풍양속에 위반되는 내용인 경우
-          <br />
-          라. 지적재산권 등 기타 타인의 권리를 침해하는 내용인 경우
-          <br />
-          마. 회사의 서비스 내용과 합치하지 않는 경우
-          <br />
-          바. 기타 관계법령에 위반된다고 판단되는 경우
-          <br />
-          제 14조 (게시물의 저작권)
-          <br />
-          1. 서비스에 게재된 자료에 대한 권리는 다음 각 호와 같습니다.
-          <br />
-          가. 회사가 작성한 저작물에 대한 저작권 기타 지적재산권은 회사에
-          귀속합니다.
-          <br />
-          나. 이용자는 회사의 서비스를 이용함으로써 얻은 정보를 회사의 사전승낙
-          없이 복제, 전송, 출판, 배포, 방송 기타 방법에 의하여 영리목적으로
-          이용하거나 제3자에게 이용하게 하여서는 안됩니다.
-          <br />
-          제 15조 (서비스 이용시간)
-          <br />
-          1. 서비스 운영은 회사의 업무상 또는 기술상 문제가 없는 한 연중무휴 1일
-          24시간을 원칙으로 합니다. 다만, 정기 점검 등의 필요로 회사가 정한
-          시간은 그러하지 아니하며 사전에 공지합니다.
-          <br />
-          2. 회사는 일부 서비스의 이용가능 시간을 별도로 정할 수 있습니다. 이
-          경우 서비스 이용안내를 통해 공지합니다.
-          <br />
-          제 16조 (서비스 제공의 중지)
-          <br />
-          1. 회사는 다음 각 호에 해당하는 경우 서비스 제공을 중지할 수 있습니다.
-          <br />
-          가. 서비스 설비의 증설, 보수 등 공사로 인한 경우
-          <br />
-          나. 전기통신사업법에 규정된 기간통신사업자가 전기통신 서비스를
-          중지했을 경우
-          <br />
-          다. 기타 불가항력적인 경우 천재지변, 국가비상사태 등의 불가항력적인
-          경우를 제외하고는 서비스 제공 중지 전에 회원에게 공지합니다.
-          <br />
-          <br />
-          제 5 장 계약해지 및 이용제한
-          <br />
-          제 17조 (계약해지 및 이용제한)
-          <br />
-          1. 회원이 이용계약을 해지하고자 하는 때에는 회원 본인이 온라인을
-          통하여 회사에 해지 신청을 하여야 합니다
-          <br />
-          2. 회사는 회원이 다음 각 호에 해당하는 행위를 했을 경우 사전통지 없이
-          이용계약을 해지하거나 서비스 이용을 제한할 수 있습니다.
-          <br />
-          가. 회사의 서비스 운영을 고의로 방해한 경우
-          <br />
-          나. 공공질서 및 미풍양속에 저해되는 내용을 유포시킨 경우
-          <br />
-          다. 타인의 이용자번호(ID) 및 비밀번호를 도용한 경우
-          <br />
-          라. 회원 가입시 허위 내용을 사용하여 신청하였을 경우
-          <br />
-          마. 타인의 명예를 손상시키거나 불이익을 주는 행위를 했을 경우
-          <br />
-          바. 컴퓨터 바이러스 프로그램 등을 유포하는 경우
-          <br />
-          사. 기타 법률에 위반되는 행위를 할 경우
-          <br />
-          <br />
-          제 6 장 개인정보 보호
-          <br />
-          제 18조 (개인정보보호정책)
-          <br />
-          1. 회사는 회원이 가입시 기재한 개인정보 및 서비스를 제공하면서 획득한
-          내용을 보호하기 위해 노력합니다. 회원의 개인정보보호에 관하여는 회사의
-          &apos;개인정보보호정책&apos; 및 관련법령이 정한 바에 의합니다.
-          <br />
-          <br />
-          제 7 장 손해 배상
-          <br />
-          제 19조 (손해 배상)
-          <br />
-          1. 회사는 회원의 부주의로 발생한 어떠한 손해에 관하여도 책임을 지지
-          않습니다.
-          <br />
-          제 20조 (면책)
-          <br />
-          1. 회사는 회원이 서비스를 이용하여 기대하는 이익을 얻지 못하였거나
-          서비스 자료에 대한 이용으로 발생하는 손해에 대해서는 책임을 지지
-          않습니다.
-          <br />
-          2. 회사는 회원의 귀책사유로 인한 서비스 이용의 장애에 대하여 책임을
-          지지 않습니다.
-          <br />
-          3. 회사는 천재지변 등의 불가항력으로 인해 서비스를 제공할 수 없는
-          경우에는 서비스 제공에 대한 책임을 지지 않습니다.
-          <br />
-          4. 회사는 모든 서비스의 자료 보관 및 전송에 관한 책임이 없으며 자료의
-          손실이 있는 경우에도 책임을 지지 않습니다.
-          <br />
-          5. 회사는 제휴사이트가 전송, 제공한 자료의 신뢰도, 정확성 등 내용에
-          관하여는 책임을 지지 않습니다.
-          <br />
-          <br />
-          제 8 장 관할법원
-          <br />
-          제 21조 (관할법원)
-          <br />
-          1. 회사와 이용자간의 콘텐츠 이용계약에 관한 소송은 제소 당시의
-          이용자의 주소에 의하고, 주소가 없는 경우 거소를 관할하는 지방법원의
-          전속관할로 합니다.
-          <br />
-          2. 제소 당시 이용자의 주소 또는 거소가 분명하지 아니한 경우에 회사와
-          이용자간의 콘텐츠 이용계약에 관한 소송에 관한 관할법원은
-          「민사소송법」에 따라 정합니다.
-          <br />
-          <br />
-          부칙
-          <br />
-          (시행일) 본 약관은 2023년 9월 1일부터 시행합니다.
-        </p>
+        <TermsData />
       </div>
     </>
   );
